@@ -20,12 +20,13 @@
     "moment": "^2.29.4",
     "node-fetch": "^3.2.3",
     "node-polyfill-webpack-plugin": "^2.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
     "react-markdown": "^8.0.2",
     "react-new-window": "^0.2.1",
+    "react-resizable-panels": "^0.0.36",
     "react-router-dom": "6",
     "react-scripts": "5.0.0",
     "rehype": "^12.0.1",
