@@ -10,7 +10,7 @@ export function SecondaryBasicInfo() {
   const { activeCard, updateActiveCard } = useCardStorage();
 
   return (
-    <>
+    <Form>
       <Form.Item label={"Variant"}>
         <Select
           value={activeCard.variant || "secondary"}
@@ -27,24 +27,28 @@ export function SecondaryBasicInfo() {
         </Select>
       </Form.Item>
       {activeCard.variant === "custom" && (
-        <>
-          <Form.Item label={"Custom height (cm)"}>
-            <Input
-              type={"text"}
-              placeholder={"Height in centimeters"}
-              value={activeCard.height}
-              onChange={(e) => updateActiveCard({ ...activeCard, height: e.target.value })}
-            />
-          </Form.Item>
-          <Form.Item label={"Custom width (cm)"}>
-            <Input
-              type={"text"}
-              placeholder={"Width in centimeters"}
-              value={activeCard.width}
-              onChange={(e) => updateActiveCard({ ...activeCard, width: e.target.value })}
-            />
-          </Form.Item>
-        </>
+        <Row gutter={8}>
+          <Col span={12}>
+            <Form.Item label={"Height (cm)"}>
+              <Input
+                type={"text"}
+                placeholder={"Height in centimeters"}
+                value={activeCard.height}
+                onChange={(e) => updateActiveCard({ ...activeCard, height: e.target.value })}
+              />
+            </Form.Item>
+          </Col>
+          <Col span={12}>
+            <Form.Item label={"Width (cm)"}>
+              <Input
+                type={"text"}
+                placeholder={"Width in centimeters"}
+                value={activeCard.width}
+                onChange={(e) => updateActiveCard({ ...activeCard, width: e.target.value })}
+              />
+            </Form.Item>
+          </Col>
+        </Row>
       )}
       <Form.Item label={"Name"}>
         <Input
@@ -116,6 +120,6 @@ export function SecondaryBasicInfo() {
           </Col>
         </Row>
       </Card>
-    </>
+    </Form>
   );
 }
