@@ -9,7 +9,7 @@ export function UnitBasicInfo() {
   const { activeCard, updateActiveCard } = useCardStorage();
 
   return (
-    <>
+    <Form>
       <Form.Item label={"Variant"}>
         <Select
           value={activeCard.variant || "card"}
@@ -25,24 +25,28 @@ export function UnitBasicInfo() {
         </Select>
       </Form.Item>
       {activeCard.variant === "custom" && (
-        <>
-          <Form.Item label={"Custom height (cm)"}>
-            <Input
-              type={"text"}
-              placeholder={"Height in centimeters"}
-              value={activeCard.height}
-              onChange={(e) => updateActiveCard({ ...activeCard, height: e.target.value })}
-            />
-          </Form.Item>
-          <Form.Item label={"Custom width (cm)"}>
-            <Input
-              type={"text"}
-              placeholder={"Width in centimeters"}
-              value={activeCard.width}
-              onChange={(e) => updateActiveCard({ ...activeCard, width: e.target.value })}
-            />
-          </Form.Item>
-        </>
+        <Row gutter={8}>
+          <Col span={12}>
+            <Form.Item label={"Height (cm)"}>
+              <Input
+                type={"text"}
+                placeholder={"Height in centimeters"}
+                value={activeCard.height}
+                onChange={(e) => updateActiveCard({ ...activeCard, height: e.target.value })}
+              />
+            </Form.Item>
+          </Col>
+          <Col span={12}>
+            <Form.Item label={"Width (cm)"}>
+              <Input
+                type={"text"}
+                placeholder={"Width in centimeters"}
+                value={activeCard.width}
+                onChange={(e) => updateActiveCard({ ...activeCard, width: e.target.value })}
+              />
+            </Form.Item>
+          </Col>
+        </Row>
       )}
       <Form.Item label={"Icons"}>
         <Select
@@ -79,6 +83,6 @@ export function UnitBasicInfo() {
           onChange={(value) => updateActiveCard({ ...activeCard, background: value })}
         />
       </Form.Item>
-    </>
+    </Form>
   );
 }
